Refactor reference check
