+++ tests/perf/__init__.py
@@ -1,4 +1,4 @@
 """OTX Perfomance tests."""
 
-# Copyright (C) 2023 Intel Corporation
+# Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
