+++ tests/perf/test_instance_segmentation.py
@@ -53,13 +53,13 @@ class TestPerfInstanceSegmentation:
 
     @pytest.mark.parametrize("fxt_model_id", MODEL_TEMPLATES, ids=MODEL_IDS, indirect=True)
     @pytest.mark.parametrize("fxt_benchmark", BENCHMARK_CONFIGS.items(), ids=BENCHMARK_CONFIGS.keys(), indirect=True)
-    def test_perf(self, fxt_model_id: str, fxt_benchmark: OTXBenchmark, fxt_check_benchmark_result: Callable):
+    def test_perf(self, fxt_model_id: str, fxt_benchmark: OTXBenchmark):
         """Benchmark performance metrics."""
         result = fxt_benchmark.run(model_id=fxt_model_id)
-        fxt_check_benchmark_result(
+        fxt_benchmark.check(
             result,
             key=(fxt_benchmark.tags["task"], fxt_benchmark.tags["data_size"], fxt_model_id),
-            checks=[
+            criteria=[
                 {
                     "name": "f-measure(train)",
                     "op": ">",
@@ -153,13 +153,13 @@ class TestPerfTilingInstanceSegmentation:
 
     @pytest.mark.parametrize("fxt_model_id", MODEL_TEMPLATES, ids=MODEL_IDS, indirect=True)
     @pytest.mark.parametrize("fxt_benchmark", BENCHMARK_CONFIGS.items(), ids=BENCHMARK_CONFIGS.keys(), indirect=True)
-    def test_perf(self, fxt_model_id: str, fxt_benchmark: OTXBenchmark, fxt_check_benchmark_result: Callable):
+    def test_perf(self, fxt_model_id: str, fxt_benchmark: OTXBenchmark):
         """Benchmark performance metrics."""
         result = fxt_benchmark.run(model_id=fxt_model_id)
-        fxt_check_benchmark_result(
+        fxt_benchmark.check(
             result,
             key=(fxt_benchmark.tags["task"], fxt_benchmark.tags["data_size"], fxt_model_id),
-            checks=[
+            criteria=[
                 {
                     "name": "f-measure(train)",
                     "op": ">",
