+++ tests/pytest.ini
@@ -5,3 +5,4 @@ markers =
   components
   reqids
   unit
+python_files = tests/**/*.py
